+++ SystemFinder/Logic/CampaignIO/Readers/SavedObjectReader.cs
@@ -1,7 +1,7 @@
 ﻿using System.Xml.Linq;
 using SystemFinder.Model.Data;
 
-namespace SystemFinder.Logic.CampaignIO
+namespace SystemFinder.Logic.CampaignIO.Readers
 {
     public class SavedObjectReader(LocationTokenReader locReader, CentReader centReader,
         FleetReader fltReader, PlanetReader plntReader)
