+++ SystemFinder/Logic/CampaignIO/Readers/ScriptReader.cs
@@ -1,9 +1,9 @@
 ﻿using System.Xml.Linq;
 using SystemFinder.Model.Data;
 
-namespace SystemFinder.Logic.CampaignIO
+namespace SystemFinder.Logic.CampaignIO.Readers
 {
-    public class cL_Reader(SavedObjectReader osavedReader)
+    public class ScriptReader
     {
         public void Read(XElement current, GalaxyData data)
         {
